+++ packages/std/src/math/uint256.rs
@@ -251,6 +251,10 @@ impl Uint256 {
             .ok_or_else(|| DivideByZeroError::new(self))
     }
 
+    pub fn checked_div_euclid(self, other: Self) -> Result<Self, DivideByZeroError> {
+        self.checked_div(other)
+    }
+
     pub fn checked_rem(self, other: Self) -> Result<Self, DivideByZeroError> {
         self.0
             .checked_rem(other.0)
@@ -1454,6 +1458,18 @@ mod tests {
         ));
         assert_eq!(
             Uint256::from(6u32).checked_div(Uint256::from(2u32)),
+            Ok(Uint256::from(3u32)),
+        );
+        assert!(matches!(
+            Uint256::MAX.checked_div_euclid(Uint256::from(0u32)),
+            Err(DivideByZeroError { .. })
+        ));
+        assert_eq!(
+            Uint256::from(6u32).checked_div_euclid(Uint256::from(2u32)),
+            Ok(Uint256::from(3u32)),
+        );
+        assert_eq!(
+            Uint256::from(7u32).checked_div_euclid(Uint256::from(2u32)),
             Ok(Uint256::from(3u32)),
         );
         assert!(matches!(