+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to
 - cosmwasm-std: Implement `saturating_add`/`sub`/`mul` for
   `Decimal`/`Decimal256`.
 - cosmwasm-std: Implement `MIN` const value for all `Uint` and `Decimal` types
+- cosmwasm-std: Implement `checked_div_euclid` for `Uint256`/`Uint512`
 - cosmwasm-std: Add `QuerierWrapper::query_wasm_contract_info` - this is just a
   convenience helper for querying `WasmQuery::ContractInfo`.